+++ routes/renderRoutes.js
@@ -1,36 +1,33 @@
 const express = require(`express`);
 const router = express.Router();
+const { getUserByID, getUserByEmail } = require("./routeHelpers");
 
 module.exports = db => {
-  router.getUserByEmail = ({ db, email }) => {
-    let queryParams = [];
-    let queryString = `SELECT * FROM users `;
-
-    queryParams.push(email);
-    queryString += `WHERE users.email = $${queryParams.length};`;
-
-    return db
-      .query(queryString, queryParams)
-      .then(data => {
-        return data.rows[0];
-      })
-      .catch(err => {
-        return { status: 500, error: err.emssage };
-      });
-  };
-
   router.get("/", (req, res) => {
-    const currentUser = req.cookies && req.cookies.user ? req.cookies.user : null;
+    const currentUser = req.cookies && req.cookies.userID ? req.cookies.userID : null;
     let user;
 
-    return router.getUserByEmail({ db, email: currentUser }).then(data => {
+    return getUserByID(db, { id: currentUser }).then(data => {
       user = data;
       res.render("index", { currentUser, user });
     });
   });
 
   router.get("/login", (req, res) => {
-    res.render("login");
+    res.render("index");
+  });
+
+  router.get("/maps/:id", (req, res) => {
+    res.render("single-map");
+  });
+
+  router.get("/users/:id", (req, res) => {
+    const currentUser = req.cookies && req.cookies.userID ? req.cookies.userID : null;
+    // let user;
+    return getUserByID(db, { id: req.params.id }).then(user => {
+      res.render("profile", { currentUser, user });
+      return user;
+    });
   });
 
   // Get user given a user email address
@@ -45,7 +42,7 @@ module.exports = db => {
       .query(queryString, queryParams)
       .then(data => {
         const result = data.rows[0];
-        res.cookie("user", data.rows[0].email);
+        res.cookie("userID", data.rows[0].id);
         res.render("index", { user: { ...result, currentUser: req.cookies.user } });
       })
       .catch(err => {
