+++ views/index.ejs
@@ -5,7 +5,6 @@
 <%- include('partials/_headnav') %>
 
 <main>
-  <%- JSON.stringify(locals) %>
   <div class="bg-gray full-width">
     <div class="hero hero-lg site-hero">
       <div class="hero-body">
