+++ views/partials/_site-head.ejs
@@ -18,8 +18,7 @@
     />
 
     <link rel="stylesheet" href="/styles/main.css" type="text/css" />
-    
   </head>
   <body>
     <div class="wrapper">
-
+      <%- JSON.stringify(locals) %>
