+++ views/profile.ejs
@@ -6,7 +6,7 @@
 
 <main>
   <div class="content-header">
-    <h1>[User Name]</h1>
+    <h1><%= user.fullname %></h1>
 
     <div class="form-group text-right">
       <select class="form-select profile-view" id="profile-view-select">
@@ -17,21 +17,7 @@
     </div>
   </div>
 
-  <section id="user-container" class="grid-container">
-    <!-- <h2>My Maps</h2>
-
-    <div class="card-grid">
-      <div class="card s-rounded">
-        <div class="card-image">
-          <img class="img-responsive" src="https://picsum.photos/800/500" />
-        </div>
-        <div class="card-header">
-          <div class="card-title h5">Map Title</div>
-          <p>999 &hearts;</p>
-        </div>
-      </div>
-    </div> -->
-  </section>
+  <section id="user-container" class="grid-container"></section>
 </main>
 
 <%- include('partials/_site-foot') %>
